Here, I have created two different notebooks for this project.
In this project, I have used the OpenCV packages and PIL library. The reason is OpenCV is a program library primarily designed for computer vision in real-time. It is a library used for the processing of images in simple languages. It is extremely efficient and enables image processing in real-time.

For computer vision projects there are many alternatives packages. Ex:- dlib, imutils, Keras, etc. 

These alternatives packages have not to give real-time computer vision. That's why I have used the OpenCV package.

Steps:-  
1. Importing the cv2
2. Importing Matplotlib
3 Using PIL library